+++ pages/index.js
@@ -2,7 +2,7 @@ import { useEffect, useState } from "react"
 import {Box} from "@mui/system"
 import { Autocomplete, TextField } from "@mui/material"
 import BookList from "../commponents/BookList"
-import SearchForm from "../commponents/SearchForm"
+import SearchField from "../commponents/SearchField"
 import { useRouter } from "next/router";
 import TabNavigation from "../commponents/TabNavigation.jsx"
 import SimpleBottomNavigation from "../commponents/BottomNavigation"
@@ -16,7 +16,7 @@ export default function Home() {
   },[])
   return (
     <Box >
-      <SearchForm/>
+      <SearchField/>
       <TabNavigation/>
       <BookList keyword={router.query.keyword}/>
       <SimpleBottomNavigation/>
