+++ styles/globals.css
@@ -156,7 +156,6 @@ height: 200px;
 .book-favorite_button {
   position:absolute;
   padding: 10px 10px;
-  border-radius: 5px;
   bottom:5px;
   left:70%;
